Fixed Zobrist hash & detection of repetition stalemate with it

+++ src/bitboards.hpp
@@ -19,7 +19,7 @@ struct GameState
     int en_passant_square = -1;
     int halfmove_clock = 0;
     int fullmove_number = 0;
-    int zobrist_hash = INT64_MAX;
+    uint64_t zobrist_hash = UINT64_MAX;
 };
